Separate interface/implementation in util/.
