+++ src/SConscript
@@ -122,6 +122,8 @@ model/network-reduction-tree.cpp
 model/network-simple-multicast.cpp
 model/sparse-optimization-info.cpp
 model/sparse-optimization-parser.cpp
+util/banner.cpp
+util/args.cpp
 util/numeric.cpp
 util/map2d.cpp
 util/accelergy_interface.cpp
