+++ src/util/banner.cpp
@@ -25,9 +25,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#pragma once
+#include "util/banner.hpp"
 
-#include <vector>
-#include <string>
-
-extern const std::vector<std::string> banner;
+const std::vector<std::string> banner =
+{
+  "  _______                __                ",
+  " /_  __(_)___ ___  ___  / /___  ____  ____ ",
+  "  / / / / __ `__ \\/ _ \\/ / __ \\/ __ \\/ __ \\",
+  " / / / / / / / / /  __/ / /_/ / /_/ / /_/ /",
+  "/_/ /_/_/ /_/ /_/\\___/_/\\____/\\____/ .___/ ",
+  "                                  /_/      "
+};
